refactored to use test.step
