+++ playwright-e2e/tests/daily-cause-list-report-test.spec.ts
@@ -1,205 +1,193 @@
 import { expect, test } from "../fixtures";
 import { config } from "../utils";
-import {SessionBookingPage} from "../page-objects/pages/hearings/session-booking.po.ts";
+import { SessionBookingPage } from "../page-objects/pages/hearings/session-booking.po.ts";
 import {
-    CaseDetailsPage,
-    CaseSearchPage,
-    HearingSchedulePage,
-    HomePage,
+  CaseDetailsPage,
+  CaseSearchPage,
+  HearingSchedulePage,
+  HomePage,
 } from "../page-objects/pages";
 
 test.describe("Daily Cause List Report tests @daily-cause-list-tests", () => {
-    test.slow();
-    // test.describe.configure({mode: "serial"});
-    //
-    // test.beforeEach(
-    //     async ({
-    //                page,
-    //                loginPage,
-    //                addNewCasePage,
-    //                caseSearchPage,
-    //                editNewCasePage,
-    //                hearingSchedulePage,
-    //            }) => {
-    //         await page.goto(config.urls.baseUrl);
-    //         await loginPage.login(config.users.testUser);
-    //         //empties cart if there is anything present
-    //         await hearingSchedulePage.sidebarComponent.emptyCaseCart();
-    //         //search for the case
-    //         await addNewCasePage.sidebarComponent.openSearchCasePage();
-    //         await caseSearchPage.searchCase(process.env.HMCTS_CASE_NUMBER as string);
-    //         await expect(editNewCasePage.caseNameField).toHaveText(
-    //             process.env.CASE_NAME as string,
-    //         );
-    //     },
-    // );
+  test.slow();
+  // test.describe.configure({mode: "serial"});
+  //
+  // test.beforeEach(
+  //     async ({
+  //                page,
+  //                loginPage,
+  //                addNewCasePage,
+  //                caseSearchPage,
+  //                editNewCasePage,
+  //                hearingSchedulePage,
+  //            }) => {
+  //         await page.goto(config.urls.baseUrl);
+  //         await loginPage.login(config.users.testUser);
+  //         //empties cart if there is anything present
+  //         await hearingSchedulePage.sidebarComponent.emptyCaseCart();
+  //         //search for the case
+  //         await addNewCasePage.sidebarComponent.openSearchCasePage();
+  //         await caseSearchPage.searchCase(process.env.HMCTS_CASE_NUMBER as string);
+  //         await expect(editNewCasePage.caseNameField).toHaveText(
+  //             process.env.CASE_NAME as string,
+  //         );
+  //     },
+  // );
 
+  test.only('List "Released" session and Generate report via reports menu', async ({
+    sessionBookingPage,
+    caseSearchPage,
+    caseDetailsPage,
+    hearingSchedulePage,
+    homePage,
+    viewReportsPage,
+    dataUtils,
+    page,
+    loginPage,
+    addNewCasePage,
+    editNewCasePage,
+  }) => {
+    await page.goto(config.urls.baseUrl);
+    await loginPage.login(config.users.testUser);
+    //empties cart if there is anything present
+    await hearingSchedulePage.sidebarComponent.emptyCaseCart();
+    //search for the case
+    await addNewCasePage.sidebarComponent.openSearchCasePage();
+    await caseSearchPage.searchCase(process.env.HMCTS_CASE_NUMBER as string);
+    await expect(editNewCasePage.caseNameField).toHaveText(
+      process.env.CASE_NAME as string,
+    );
 
-    test.only('List "Released" session and Generate report via reports menu', async ({
-                                                                                         sessionBookingPage,
-                                                                                         caseSearchPage,
-                                                                                         caseDetailsPage,
-                                                                                         hearingSchedulePage,
-                                                                                         homePage,
-                                                                                         viewReportsPage,
-                                                                                         dataUtils,
-                                                                                         page,
-                                                                                         loginPage,
-                                                                                         addNewCasePage,
-                                                                                         editNewCasePage
-                                                                                     }) => {
+    await sessionBookingPage.sidebarComponent.openHearingSchedulePage();
 
-        await page.goto(config.urls.baseUrl);
-        await loginPage.login(config.users.testUser);
-        //empties cart if there is anything present
-        await hearingSchedulePage.sidebarComponent.emptyCaseCart();
-        //search for the case
-        await addNewCasePage.sidebarComponent.openSearchCasePage();
-        await caseSearchPage.searchCase(process.env.HMCTS_CASE_NUMBER as string);
-        await expect(editNewCasePage.caseNameField).toHaveText(
-            process.env.CASE_NAME as string,
-        );
+    await sessionBookingPage.updateAdvancedFilterConfig(
+      sessionBookingPage.CONSTANTS.CASE_LISTING_REGION_WALES,
+      sessionBookingPage.CONSTANTS
+        .CASE_LISTING_CLUSTER_WALES_CIVIL_FAMILY_TRIBUNALS,
+      sessionBookingPage.CONSTANTS
+        .CASE_LISTING_LOCALITY_PONTYPRIDD_COUNTY_COURT,
+      sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
+    );
 
-        await sessionBookingPage.sidebarComponent.openHearingSchedulePage();
+    await hearingSchedulePage.clearDownSchedule(
+      sessionBookingPage.CONSTANTS.SESSION_DETAILS_CANCELLATION_CODE_CANCEL,
+      sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
+      dataUtils.generateDateInDdMmYyyyWithHypenSeparators(0),
+    );
 
-        await sessionBookingPage.updateAdvancedFilterConfig(
-            sessionBookingPage.CONSTANTS.CASE_LISTING_REGION_WALES,
-            sessionBookingPage.CONSTANTS
-                .CASE_LISTING_CLUSTER_WALES_CIVIL_FAMILY_TRIBUNALS,
-            sessionBookingPage.CONSTANTS
-                .CASE_LISTING_LOCALITY_PONTYPRIDD_COUNTY_COURT,
-            sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
-        );
+    // Test data
+    const roomData = {
+      roomName:
+        sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
+      column: sessionBookingPage.CONSTANTS.CASE_LISTING_COLUMN_ONE,
+      caseNumber: process.env.HMCTS_CASE_NUMBER as string,
+      sessionDuration:
+        sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_DURATION_1_00,
+      hearingType:
+        sessionBookingPage.CONSTANTS.CASE_LISTING_HEARING_TYPE_APPLICATION,
+      cancelReason:
+        sessionBookingPage.CONSTANTS.CASE_LISTING_CANCEL_REASON_AMEND,
+    };
 
-        await hearingSchedulePage.clearDownSchedule(
-            sessionBookingPage.CONSTANTS.SESSION_DETAILS_CANCELLATION_CODE_CANCEL,
-            sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
-            dataUtils.generateDateInDdMmYyyyWithHypenSeparators(0),
-        );
+    await createHearingSession(
+      roomData.caseNumber,
+      homePage,
+      caseSearchPage,
+      caseDetailsPage,
+      hearingSchedulePage,
+      roomData,
+      sessionBookingPage,
+    );
 
-        // Test data
-        const roomData = {
-            roomName:
-            sessionBookingPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
-            column: sessionBookingPage.CONSTANTS.CASE_LISTING_COLUMN_ONE,
-            caseNumber: process.env.HMCTS_CASE_NUMBER as string,
-            sessionDuration:
-            sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_DURATION_1_00,
-            hearingType:
-            sessionBookingPage.CONSTANTS.CASE_LISTING_HEARING_TYPE_APPLICATION,
-            cancelReason:
-            sessionBookingPage.CONSTANTS.CASE_LISTING_CANCEL_REASON_AMEND,
-        };
+    const reportData = {
+      //numeric, current day of the month
+      todayDate: dataUtils.generateDateInYyyyMmDdNoSeparators(0),
+      locality:
+        viewReportsPage.CONSTANTS.CASE_LISTING_LOCALITY_PONTYPRIDD_COUNTY_COURT,
+      location:
+        viewReportsPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
+      jurisdiction: viewReportsPage.CONSTANTS.JURISDICTION_CIVIL,
+      service: viewReportsPage.CONSTANTS.SERVICE_DAMAGES,
+    };
 
-        await createHearingSession(
-            roomData.caseNumber,
-            homePage,
-            caseSearchPage,
-            caseDetailsPage,
-            hearingSchedulePage,
-            roomData,
-            sessionBookingPage,
-        );
+    //open reports menu and check generated External hearing list report
+    await viewReportsPage.reportRequestPageActions(
+      reportData.todayDate,
+      reportData.locality,
+      reportData.location,
+      reportData.jurisdiction,
+      dataUtils.getFormattedDateForReportAssertion(),
+    );
 
-        const reportData = {
-            //numeric, current day of the month
-            todayDate: dataUtils.generateDateInYyyyMmDdNoSeparators(0),
-            locality:
-            viewReportsPage.CONSTANTS.CASE_LISTING_LOCALITY_PONTYPRIDD_COUNTY_COURT,
-            location:
-            viewReportsPage.CONSTANTS.CASE_LISTING_LOCATION_PONTYPRIDD_CRTRM_1,
-            jurisdiction: viewReportsPage.CONSTANTS.JURISDICTION_CIVIL,
-            service: viewReportsPage.CONSTANTS.SERVICE_DAMAGES,
+    // open reports menu and check generated Internal hearing list report
+    await viewReportsPage.reportRequestPageActions(
+      reportData.todayDate,
+      reportData.locality,
+      reportData.location,
+      reportData.jurisdiction,
+      dataUtils.getFormattedDateForReportAssertion(),
+      reportData.service,
+    );
+  });
 
-        };
+  async function createHearingSession(
+    caseName: string,
+    homePage: HomePage,
+    caseSearchPage: CaseSearchPage,
+    caseDetailsPage: CaseDetailsPage,
+    hearingSchedulePage: HearingSchedulePage,
+    roomData: {
+      roomName: string;
+      column: string;
+      caseNumber: string;
+      sessionDuration: string;
+      hearingType: string;
+      cancelReason: string;
+    },
+    sessionBookingPage: SessionBookingPage,
+  ) {
+    // Check if the close case button in upper bar is present
+    await expect(homePage.upperbarComponent.closeCaseButton).toBeVisible();
+    //check current case drop down menu in upper bar
+    await expect(
+      homePage.upperbarComponent.currentCaseDropdownButton,
+    ).toBeVisible();
+    await homePage.upperbarComponent.currentCaseDropdownButton.click();
 
-        //open reports menu and check generated External hearing list report
-        await viewReportsPage.reportRequestPageActions(
-            reportData.todayDate,
-            reportData.locality,
-            reportData.location,
-            reportData.jurisdiction,
-            dataUtils.getFormattedDateForReportAssertion(),
-        );
+    await expect(
+      homePage.upperbarComponent.currentCaseDropdownList,
+    ).toContainText(homePage.upperbarComponent.currentCaseDropDownItems);
 
-        // open reports menu and check generated Internal hearing list report
-        await viewReportsPage.reportRequestPageActions(
-            reportData.todayDate,
-            reportData.locality,
-            reportData.location,
-            reportData.jurisdiction,
-            dataUtils.getFormattedDateForReportAssertion(),
-            reportData.service,
-        );
-    });
+    //add case to cart
+    await caseSearchPage.sidebarComponent.openSearchCasePage();
+    await caseSearchPage.searchCase(caseName);
 
-        async function createHearingSession(
-            caseName: string,
-            homePage: HomePage,
-            caseSearchPage: CaseSearchPage,
-            caseDetailsPage: CaseDetailsPage,
-            hearingSchedulePage: HearingSchedulePage,
-            roomData: {
-                roomName: string;
-                column: string;
-                caseNumber: string;
-                sessionDuration: string;
-                hearingType: string;
-                cancelReason: string;
-            },
-            sessionBookingPage: SessionBookingPage,
-        ) {
-            // Check if the close case button in upper bar is present
-            await expect(homePage.upperbarComponent.closeCaseButton).toBeVisible();
-            //check current case drop down menu in upper bar
-            await expect(
-                homePage.upperbarComponent.currentCaseDropdownButton,
-            ).toBeVisible();
-            await homePage.upperbarComponent.currentCaseDropdownButton.click();
+    await expect(caseDetailsPage.addToCartButton).toBeVisible();
+    await caseDetailsPage.addToCartButton.click();
+    await expect(caseDetailsPage.sidebarComponent.cartButton).toBeEnabled();
 
-            await expect(
-                homePage.upperbarComponent.currentCaseDropdownList,
-            ).toContainText(homePage.upperbarComponent.currentCaseDropDownItems);
+    //go to hearing schedule page
+    await expect(hearingSchedulePage.sidebarComponent.sidebar).toBeVisible();
+    await hearingSchedulePage.sidebarComponent.openHearingSchedulePage();
 
-            //add case to cart
-            await caseSearchPage.sidebarComponent.openSearchCasePage();
-            await caseSearchPage.searchCase(caseName);
+    //schedule hearing
+    await hearingSchedulePage.waitForLoad();
 
-            await expect(caseDetailsPage.addToCartButton).toBeVisible();
-            await caseDetailsPage.addToCartButton.click();
-            await expect(caseDetailsPage.sidebarComponent.cartButton).toBeEnabled();
+    await hearingSchedulePage.scheduleHearingWithBasket(
+      roomData.roomName,
+      roomData.column,
+      roomData.caseNumber,
+    );
 
-            //go to hearing schedule page
-            await expect(hearingSchedulePage.sidebarComponent.sidebar).toBeVisible();
-            await hearingSchedulePage.sidebarComponent.openHearingSchedulePage();
+    //session booking page
+    await sessionBookingPage.bookSession(
+      sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_DURATION_1_00,
+      sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_STATUS_TYPE_RELEASED,
+    );
 
-            //schedule hearing
-            await hearingSchedulePage.waitForLoad();
-
-            await hearingSchedulePage.scheduleHearingWithBasket(
-                roomData.roomName,
-                roomData.column,
-                roomData.caseNumber,
-            );
-
-            //session booking page
-            await sessionBookingPage.bookSession(
-                sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_DURATION_1_00,
-                sessionBookingPage.CONSTANTS.CASE_LISTING_SESSION_STATUS_TYPE_RELEASED,
-            );
-
-            //confirm listing
-            await expect(
-                hearingSchedulePage.confirmListingReleasedStatus,
-            ).toBeVisible();
-        }
-
-
-    });
-
-
-
-
-
-
-
+    //confirm listing
+    await expect(
+      hearingSchedulePage.confirmListingReleasedStatus,
+    ).toBeVisible();
+  }
+});
